images: convert-url: proxy entities images requests in federated mode
