+++ server/controllers/images/lib/convert_and_cleanup_image_url.ts
@@ -1,19 +1,24 @@
 import { cleanupImageUrl } from '#data/dataseed/dataseed'
 import { isUrl } from '#lib/boolean_validations'
 import { newError } from '#lib/error/error'
+import { signedFederatedRequest } from '#lib/federation/signed_federated_request'
 import { isPrivateUrl } from '#lib/network/is_private_url'
 import { assertString } from '#lib/utils/assert_types'
 import { log } from '#lib/utils/logs'
-import config from '#server/config'
+import config, { federatedMode } from '#server/config'
 import type { AbsoluteUrl } from '#types/common'
 import type { ImageContainer } from '#types/image'
-import convertImageUrl from './convert_image_url.js'
+import type { AuthentifiedReq } from '#types/server'
+import convertImageUrl, { type ImportedImage } from './convert_image_url.js'
 
 const { enabled: dataseedEnabled } = config.dataseed
 
-export async function convertAndCleanupImageUrl ({ container, url }: { container: ImageContainer, url: AbsoluteUrl }) {
+export async function convertAndCleanupImageUrl ({ container, url }: { container: ImageContainer, url: AbsoluteUrl }, req?: AuthentifiedReq) {
   assertString(container)
   assertString(url)
+  if (federatedMode && req && container === 'entities') {
+    return proxyRequestToConvertAndCleanupEntityImageUrl(req, url)
+  }
   const originalUrl = url
   if (dataseedEnabled && container === 'entities') {
     const res = await cleanupImageUrl(url)
@@ -33,3 +38,11 @@ const bannedHashes = new Set([
   // BNF placeholder
   '34ae223423391eeb6bcd31bf177e77c13aa013a4',
 ])
+
+async function proxyRequestToConvertAndCleanupEntityImageUrl (req: AuthentifiedReq, url: AbsoluteUrl) {
+  return signedFederatedRequest(req, 'post', '/api/images', {
+    action: 'convert-url',
+    container: 'entities',
+    url,
+  }) as Promise<Partial<ImportedImage>>
+}
